All charges working, copy ability tracking for save states broken, debug needs cleanup
