+++ src/training/character_specific/mod.rs
@@ -2,6 +2,7 @@ use smash::app::{self};
 
 mod bowser;
 pub mod items;
+pub mod kirby;
 pub mod pikmin;
 pub mod ptrainer;
 pub mod steve;
