+++ src/training/mod.rs
@@ -7,7 +7,7 @@ use crate::common::{
 use crate::hitbox_visualizer;
 use crate::input::*;
 use crate::logging::*;
-use crate::training::character_specific::{items, pikmin, ptrainer};
+use crate::training::character_specific::{items, kirby, pikmin, ptrainer};
 use skyline::hooks::{getRegionAddress, InlineCtx, Region};
 use skyline::nn::ro::LookupSymbol;
 use smash::app::{self, enSEType, lua_bind::*, utility};
@@ -963,6 +963,7 @@ pub fn training_mods() {
     ui::init();
     pikmin::init();
     ptrainer::init();
+    kirby::init();
 
     //#[cfg(debug_assertions)]
     debug::init();
